Fix signup bug and add error message for invalid username and password
Previously, when a user attempted to login with an invalid username/password, the login button would just not do anything, I added an error message so that the user knows the reason for their inability to login.
I also had a bug where even when someone signed up with a non-villanova.edu email, their user information would still be added to my database. I fixed this with an else-if.