Fix profile picture selection bug and modified editprofileviewcontroller fields to always contain most recent user info
Previously, I had a bug where I could not change the profile picture because I was using the file name provided by the phone which was too long. I changed it so that each selected image file is just named "image.png" when stored in parse. I also had it so that whenever my editprofileviewcontroller showed, the fields were automatically set to empty which was problematic. I fixed that by setting the fields to retrieved info from parse in the viewdidload in the editprofileviewcontroller.