Add info popup for info page
I used a cocoa pod to create a custom popup in the info section. The popup explains what moneyline odds are so that users can take that information into account when looking at the odds presented in my info section. The info is also useful for when the user is creating their bracket.